Some small fixes, added another variable that can be overrided by an addon (in this case the Defensio addon). $comment_meta_information, normally shows info about the datetime and IP. In defensio this variable is defined as info about datetime, IP and Spaminess.
git-svn-id: http://pixelpost.googlecode.com/svn/trunk@371 ff196f25-2e1a-0410-95ca-3d8fd8d346c5

+++ admin/comments.php
@@ -274,9 +274,16 @@ if($_GET['view'] == "comments") {
 				$admin_lang_cmnt_name <a target=\"_blank\" href=\"$url\">$name</a>
 				$admin_lang_cmnt_email $email <br />$admin_lang_cmnt_comment
 				<b>$message</b><br />
-				$admin_lang_cmnt_image: \"$imagename\"<br />
-				<i>$admin_lang_cmnt_commenter $datetime. $admin_lang_cmnt_ip  $ip.<br /></i>
-				<input type='checkbox' name='moderate_commnts_boxes[]' value='$id' />
+				$admin_lang_cmnt_image: \"$imagename\"<br />";
+		    if (isset($comment_meta_information)){
+		     // this variable can be used to overide display of the meta information for a comment
+		     // is used for the defensio addon.
+				 echo $comment_meta_information;
+				} else {
+				 //show default
+				 echo "<i>$admin_lang_cmnt_commenter $datetime. $admin_lang_cmnt_ip  $ip.<br /></i>";
+				}				
+				echo"<input type='checkbox' name='moderate_commnts_boxes[]' value='$id' />
 				<a onclick=\"return  confirmDeleteComment()\" href=\"$PHP_SELF?view=comments&amp;action=delete&amp;delid=$id\">
 				[$admin_lang_cmnt_delete]</a>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
